Change a to a_u8 and add b_u8 plus fun init_A()
This compiles and runs if this.a_u8 = 12 is NOT commented out, but
fails it is even with ponyc branch
WIP-Allow-constructors-to-call-functions-for-class-initialization.

Fails with:

    this.init_A()
               ^
    Info:
    /home/wink/prgs/pony/class-initialization/main.pony:8:5: receiver type: C tag
        this.init_A()
        ^
    /home/wink/prgs/pony/class-initialization/main.pony:11:3: target type: C ref
      fun ref init_A() =>
      ^
    /home/wink/prgs/pony/class-initialization/main.pony:8:5: C tag is not a subtype of C ref: tag is not a subcap of ref
        this.init_A()
        ^
    /home/wink/prgs/pony/class-initialization/main.pony:8:5: this might be possible if all fields were already defined
        this.init_A()
        ^

It seems like "this" in the constructor is a "tag" before initialization
is complete which would make it impossible to call functions that change
things, which makes some sense. Maybe the idea of creating "ifun" which
can only be called from constructor a solution. Although I'm going to
try to hack around it before adding a new function token.